+++ pkg/appmanager/routing.go
@@ -28,8 +28,8 @@ import (
 
 	log "github.com/F5Networks/k8s-bigip-ctlr/pkg/vlogger"
 
-	routeapi "github.com/openshift/origin/pkg/route/api"
-	"k8s.io/client-go/pkg/apis/extensions/v1beta1"
+	routeapi "github.com/openshift/api/route/v1"
+	"k8s.io/api/extensions/v1beta1"
 )
 
 const httpRedirectIRuleName = "http_redirect_irule"